Proyectos/Python IA/requirements.txt
twitchio==2.11.0
python-dotenv==1.0.0
